# openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
    FlowBaseInfo:
      type: object
      properties:
        fdr:
          type: string
        created:
          type: string
        organizationId:
          type: string
...
    LocalDate:
      type: string
      format: date
      examples:
        - 2022-03-10
...
    MultipleFlowsResponse:
      type: object
      properties:
        dateFrom:
          $ref: "#/components/schemas/LocalDate"
        dateTo:
          $ref: "#/components/schemas/LocalDate"
        data:
          type: array
          items:
            $ref: "#/components/schemas/FlowBaseInfo"
    ProblemJson:
      type: object
      properties:
        title:
          type: string
          description: "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable"
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          examples:
            - 200
          maximum: 600
          minimum: 100
        details:
          type: string
          description: A human readable explanation specific to this occurrence of the problem.
          examples:
            - There was an error processing the request
  responses:
    ErrorResponse400:
      description: Default app exception for status 400
      content:
        application/json:
          example:
            httpStatusCode: 400
            httpStatusDescription: Bad Request
            appErrorCode: FDR-XXXX
            errors:
              - path: <detail.path.if-exist>
                message: <detail.message>
          schema:
            $ref: "#/components/schemas/ProblemJson"
...
    ErrorResponse500:
      description: Internal Server Error
      content:
        application/json:
          example:
            errorId: 50905466-1881-457b-b42f-fb7b2bfb1610
            httpStatusCode: 500
            httpStatusDescription: Internal Server Error
            appErrorCode: FDR-0500
            errors:
              - message: An unexpected error has occurred. Please contact support.
          schema:
            $ref: "#/components/schemas/ProblemJson"
  securitySchemes:
    ApiKey:
      type: apiKey
      name: Ocp-Apim-Subscription-Key
      in: header
info:
  description: |
    Technical support API for FDR (aka "Flussi di Rendicontazione")

    ## OPERATIONAL ERROR CODES

    <details><summary>Details</summary>
    NAME | CODE | ERROR MESSAGE TEMPLATE | DESCRIPTION
    - | - | - | -
    **FDR-0400** | *DATE_BAD_REQUEST* | Bad request | An error occurred while validating the provided dates. <br>Date from and date to must be either null o filled <br>in, if present, date from must be less than date to
    **FDR-0404** | *FLOW_NOT_FOUND* | No flow found | An error occurred during the search of flow. The needed <br>flow does not exists in current environment
    **FDR-1404** | *FLOW_NOT_FOUND_CLIENT* | No flow found | An error occurred during the search of flow. The needed <br>flow was not found
    **FDR-1401** | *UNAUTHORIZED_CLIENT* | An unexpected error has occurred. Please contact support | An error occurred while invoking external services
    **FDR-0500** | *ERROR* | An unexpected error has occurred. Please contact support | An error occurred during computation. This could be <br>caused by an applicative error and it is probably required <br>to open an issue.
    **FDR-0400** | *BAD_REQUEST* | Bad request | A generic 'Bad Request' error is occurred during request <br>validation
    **FDR-0401** | *BAD_REQUEST_INPUT_JSON* | Bad request. Field [%] is equals to [%] but this is <br>not a valid value. | A generic error occurred during execution of request <br>syntactic validation
...
servers:
...
    variables:
...
paths:
...
  /psps/{pspId}:
    get:
      summary: This API allow to retrieve a list of FdR for a specific PSP
      description: |
        ## Description:
        Retrieves a list of FdR for a given Payment Service Provider (PSP) within a specified date range,  optionally filtered by flow name. If no dates are specified, data from the last 7 days is returned.

        ## Error codes:
        APPLICATIVE CODE | HTTP CODE | MESSAGE
        - | - | -
        **FDR-0400** | 400 | Bad request


        ## API properties:
        PROPERTY | VALUE
        - | -
        ***Internal*** | Y
        ***External*** | Y
        ***Synchronous*** | Synchronous
        ***Authorization*** | N
        ***Authentication*** | Y (Subscription Key)
        ***TPS*** | 1.0/sec
        ***Idempotency*** | Y
        ***Stateless*** | Y
        ***Read/Write Intensive*** | Read
        ***Cacheable*** | Y
      operationId: IPSPController_searchFlowByPsp
      tags:
        - PSP
      parameters:
        - description: "The PSP identifier, used as a search filter"
          example: "88888888888"
          required: true
          name: pspId
          in: path
          schema:
            type: string
        - description: The starting date for retrieving flows
          example: 2025-03-10
          name: dateFrom
          in: query
          schema:
            $ref: "#/components/schemas/LocalDate"
        - description: The ending date for retrieving flows
          example: 2025-03-10
          name: dateTo
          in: query
          schema:
            $ref: "#/components/schemas/LocalDate"
        - description: "The flow name, used as a search filter"
          example: 2025-01-0188888888888-0001
          name: flowId
          in: query
          schema:
            type: string
        - description: "The organization fiscal code, used as a search filter"
          example: "99999999999"
          name: organizationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultipleFlowsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse400"
        "500":
          $ref: "#/components/responses/ErrorResponse500"
  /psps/{pspId}/iur/{iur}:
    get:
      summary: This API allow to retrieve a list of FdR for a specific PSP and IUV
      description: |
        ## Description:
        Retrieves a list of FdR for a given Payment Service Provider (PSP) and IUV within a specified date range. If no dates are specified, data from the last 7 days is returned.

        ## Error codes:
        APPLICATIVE CODE | HTTP CODE | MESSAGE
        - | - | -
        **FDR-0400** | 400 | Bad request


        ## API properties:
        PROPERTY | VALUE
        - | -
        ***Internal*** | Y
        ***External*** | Y
        ***Synchronous*** | Synchronous
        ***Authorization*** | N
        ***Authentication*** | Y (Subscription Key)
        ***TPS*** | 1.0/sec
        ***Idempotency*** | Y
        ***Stateless*** | Y
        ***Read/Write Intensive*** | Read
        ***Cacheable*** | Y
      operationId: IPSPController_searchFlowByPspAndIur
      tags:
        - PSP
      parameters:
        - description: "The payment's IUR code, used as a search filter"
          example: "17854456582215"
          name: iur
          in: path
          required: true
          schema:
            type: string
            pattern: "^(.{1,35})$"
        - description: "The PSP identifier, used as a search filter"
          example: "88888888888"
          required: true
          name: pspId
          in: path
          schema:
            type: string
        - description: The starting date for retrieving flows
          example: 2025-03-10
          name: dateFrom
          in: query
          schema:
            $ref: "#/components/schemas/LocalDate"
        - description: The ending date for retrieving flows
          example: 2025-03-10
          name: dateTo
          in: query
          schema:
            $ref: "#/components/schemas/LocalDate"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultipleFlowsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse400"
        "500":
          $ref: "#/components/responses/ErrorResponse500"
  /psps/{pspId}/iuv/{iuv}:
    get:
      summary: This API allow to retrieve a list of FdR for a specific PSP and IUV
      description: |
        ## Description:
        Retrieves a list of FdR for a given Payment Service Provider (PSP) and IUV within a specified date range. If no dates are specified, data from the last 7 days is returned.

        ## Error codes:
        APPLICATIVE CODE | HTTP CODE | MESSAGE
        - | - | -
        **FDR-0400** | 400 | Bad request


        ## API properties:
        PROPERTY | VALUE
        - | -
        ***Internal*** | Y
        ***External*** | Y
        ***Synchronous*** | Synchronous
        ***Authorization*** | N
        ***Authentication*** | Y (Subscription Key)
        ***TPS*** | 1.0/sec
        ***Idempotency*** | Y
        ***Stateless*** | Y
        ***Read/Write Intensive*** | Read
        ***Cacheable*** | Y
      operationId: IPSPController_searchFlowByPspAndIuv
      tags:
        - PSP
      parameters:
        - description: "The payment's IUV code, used as a search filter"
          example: "17854456582215"
          name: iuv
          in: path
          required: true
          schema:
            type: string
            pattern: "^(.{1,35})$"
        - description: "The PSP identifier, used as a search filter"
          example: "88888888888"
          required: true
          name: pspId
          in: path
          schema:
            type: string
        - description: The starting date for retrieving flows
          example: 2025-03-10
          name: dateFrom
          in: query
          schema:
            $ref: "#/components/schemas/LocalDate"
        - description: The ending date for retrieving flows
          example: 2025-03-10
          name: dateTo
          in: query
          schema:
            $ref: "#/components/schemas/LocalDate"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultipleFlowsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse400"
        "500":
          $ref: "#/components/responses/ErrorResponse500"